add linter as a command to tool cli
